Desktop/uni/Semester 10/Deep Learning in computer vision/Assignment/requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0.3
PyYAML>=6.0
Pillow>=10.0
numpy>=1.26
scikit-learn>=1.4
matplotlib>=3.8